Add `torchaudio.transforms.GriffinLim()`. Replace `torchaudio.ITransform` with `IModule<Tensor, Tensor>`.
